refactor: Move  UiViewModel to HomeScreenViewModel
#222
