+++ app/src/main/java/org/nsh07/wikireader/MainActivity.kt
@@ -20,13 +20,13 @@ import org.nsh07.wikireader.data.AppStatus
 import org.nsh07.wikireader.data.isRtl
 import org.nsh07.wikireader.data.toColor
 import org.nsh07.wikireader.ui.AppScreen
+import org.nsh07.wikireader.ui.homeScreen.viewModel.HomeScreenViewModel
 import org.nsh07.wikireader.ui.settingsScreen.viewModel.SettingsViewModel
 import org.nsh07.wikireader.ui.theme.WikiReaderTheme
-import org.nsh07.wikireader.ui.viewModel.UiViewModel
 
 class MainActivity : ComponentActivity() {
 
-    val viewModel: UiViewModel by viewModels<UiViewModel>(factoryProducer = { UiViewModel.Factory })
+    val viewModel: HomeScreenViewModel by viewModels<HomeScreenViewModel>(factoryProducer = { HomeScreenViewModel.Factory })
     val settingsViewModel: SettingsViewModel by viewModels<SettingsViewModel>(factoryProducer = { SettingsViewModel.Factory })
 
     override fun onCreate(savedInstanceState: Bundle?) {
