+++ app/src/main/java/org/nsh07/wikireader/data/WRStatus.kt
@@ -9,6 +9,11 @@ enum class SavedStatus {
     NOT_SAVED, SAVING, SAVED
 }
 
+/**
+ * Enum class to track the app's startup status. On startup, the app status is
+ * [AppStatus.NOT_INITIALIZED]. After all the settings are loaded and applied, the status is set to
+ * [AppStatus.INITIALIZED]. This status is used to control the visibility of the splash screen
+ */
 enum class AppStatus {
     INITIALIZED, NOT_INITIALIZED
 }
